add data exchange and crypto fix

+++ Crypto.cpp
@@ -7,14 +7,14 @@
 using namespace std;
 
 //finding GCD of two numbers
-long long gcd(long long a, long long b) {
+int gcd(int a, int b) {
 	if (b == 0)
 		return a;
 	return gcd(b, a % b);
 }
 
 //checking a number for simplicity by iterating through the odd numbers up to sqrt(N)
-bool IsPrime(long long &N) {
+bool IsPrime(int &N) {
 	if (N == 1) {
 		return 0;
 	}
@@ -29,7 +29,7 @@ bool IsPrime(long long &N) {
 			else {
 				double n = sqrt(N);
 				int flag = 0;
-				for (long long i = 3; i <= n; i++) {
+				for (int i = 3; i <= n; i++) {
 					if (N % i == 0) {
 						return 0;
 						flag = 1;
@@ -45,8 +45,8 @@ bool IsPrime(long long &N) {
 }
 
 //Euler function (counting the number of numbers less than N and mutually prime with it)
-long long Euler(long long& N) {
-	long long sum;
+int Euler(int& N) {
+	int sum;
 	if (N == 1) {
 		sum = 1;
 	}
@@ -61,7 +61,7 @@ long long Euler(long long& N) {
 			else {
 				sum = 2;
 				if (N % 2 == 0) {
-					for (long long i = 3; i < N / 2; i = i + 2) {
+					for (int i = 3; i < N / 2; i = i + 2) {
 						if (gcd(N, i) == 1) {
 							sum = sum + 2;
 							//cout << i << " " << N - i << endl;
@@ -71,7 +71,7 @@ long long Euler(long long& N) {
 				}
 				else {
 					sum = 4;
-					for (long long i = 3; i <= N / 2; i++) {
+					for (int i = 3; i <= N / 2; i++) {
 						if (gcd(N, i) == 1) {
 							sum = sum + 2;
 							//cout << i << " " << N - i << endl;
@@ -85,18 +85,18 @@ long long Euler(long long& N) {
 }
 
 //Euclid's advanced algorithm
-long long Euclid_algo(long long a, long long b) {
+int Euclid_algo(int a, int b) {
 	if (a < 1 or b < 2) {
 		return -1;
 	}
-	long long u1 = b;
-	long long u2 = 0;
-	long long v1 = a;
-	long long v2 = 1;
+	int u1 = b;
+	int u2 = 0;
+	int v1 = a;
+	int v2 = 1;
 	while (v1 != 0) {
-		long long q = u1 / v1;
-		long long t1 = u1 - q * v1;
-		long long t2 = u2 - q * v2;
+		int q = u1 / v1;
+		int t1 = u1 - q * v1;
+		int t2 = u2 - q * v2;
 		u1 = v1;
 		u2 = v2;
 		v1 = t1;
@@ -106,7 +106,7 @@ long long Euclid_algo(long long a, long long b) {
 }
 
 //checking the keys generation capability
-bool IsGenPos(long long& p, long long& q) {
+bool IsGenPos(int& p, int& q) {
 	if (p * q >= 256) {
 		return 1;
 	}
@@ -116,12 +116,12 @@ bool IsGenPos(long long& p, long long& q) {
 }
 
 //open and private keys generation
-pair<pair<long long, long long>, long long> key_generation(long long &p, long long &q) {
-	long long N = p * q;
-	long long f = Euler(N);
-	long long e = f - 1;
-	long long d = Euclid_algo(e, f);
-	pair <pair<long long, long long>, long long> a;
+pair<pair<int, int>, int> key_generation(int &p, int &q) {
+	int N = p * q;
+	int f = Euler(N);
+	int e = f - 1;
+	int d = Euclid_algo(e, f);
+	pair <pair<int, int>, int> a;
 	a.first.first = e;
 	a.first.second = d;
 	a.second = N;
@@ -130,21 +130,21 @@ pair<pair<long long, long long>, long long> key_generation(long long &p, long lo
 }
 
 //encrypting an incoming message
-vector<long long> translation(string& message, pair <pair<long long, long long>, long long> &keys) {
-	vector<long long> trans1(message.size()), trans2(message.size());
-	for (long long i = 0; i < message.size(); i++) {
+vector<int> translation(string& message, pair <pair<int, int>, int> &keys) {
+	vector<int> trans1(message.size()), trans2(message.size());
+	for (int i = 0; i < message.size(); i++) {
 		trans1[i] = (int)message[i];
 	}
 
 	/*cout << "trans1: ";
-	for (long long i = 0; i < trans1.size(); i++) {
+	for (int i = 0; i < trans1.size(); i++) {
 		cout << trans1[i] << " ";
 	}
 	cout << endl;*/
 
-	for (long long i = 0; i < message.size(); i++) {
-		long long count = 0;
-		long long res = 1;
+	for (int i = 0; i < message.size(); i++) {
+		int count = 0;
+		int res = 1;
 		while (count < keys.first.first) {
 			res = (res * trans1[i]) % keys.second;
 			count++;
@@ -152,7 +152,7 @@ vector<long long> translation(string& message, pair <pair<long long, long long>,
 		trans2[i] = res;
 	}
 	/*cout << "trans2: ";
-	for (long long i = 0; i < trans2.size(); i++) {
+	for (int i = 0; i < trans2.size(); i++) {
 		cout << trans2[i] << " ";
 	}
 	cout << endl;*/
@@ -161,12 +161,12 @@ vector<long long> translation(string& message, pair <pair<long long, long long>,
 }
 
 //decryption and assembly of the final message
-string retranslation(vector<long long> &server_gift, pair <pair<long long, long long>, long long> &keys) {
-	vector<long long> gift(server_gift.size());
+string retranslation(vector<int> &server_gift, pair <pair<int, int>, int> &keys) {
+	vector<int> gift(server_gift.size());
 	string answer;
-	for (long long i = 0; i < server_gift.size(); i++) {
-		long long count = 0;
-		long long res = 1;
+	for (int i = 0; i < server_gift.size(); i++) {
+		int count = 0;
+		int res = 1;
 		while (count < keys.first.second) {
 			res = (res * server_gift[i]) % keys.second;
 			count++;
@@ -176,7 +176,7 @@ string retranslation(vector<long long> &server_gift, pair <pair<long long, long 
 	}
 
 	/*cout << "gift: ";
-	for (long long i = 0; i < server_gift.size(); i++) {
+	for (int i = 0; i < server_gift.size(); i++) {
 		cout << gift[i] << " ";
 	}
 	cout << endl;*/
@@ -186,7 +186,7 @@ string retranslation(vector<long long> &server_gift, pair <pair<long long, long 
 
 /*int main() {
 	cout << "Введите два простых числа p и q:" << endl;
-	long long p, q;
+	int p, q;
 	cin >> p >> q;
 	int a = IsGenPos(p, q);
 
@@ -195,10 +195,10 @@ string retranslation(vector<long long> &server_gift, pair <pair<long long, long 
 		cin >> p >> q;
 		a = IsGenPos(p, q);
 	}
-	pair<pair<long long, long long>, long long> work = key_generation(p, q);
+	pair<pair<int, int>, int> work = key_generation(p, q);
 	string input;
 	getline(cin, input);
-	vector<long long> to_server = translation(input, work);
+	vector<int> to_server = translation(input, work);
 	string result = retranslation(to_server, work);
 	cout << result;
 	return 0;
